+++ src/assets/ba_data/python/babase/_hooks.py
@@ -342,7 +342,8 @@ def hash_strings(inputs: list[str]) -> str:
 
 def have_account_v2_credentials() -> bool:
     """Do we have primary account-v2 credentials set?"""
-    return _babase.app.accounts.have_primary_credentials()
+    assert _babase.app.plus is not None
+    return _babase.app.plus.accounts.have_primary_credentials()
 
 
 def implicit_sign_in(
@@ -351,7 +352,8 @@ def implicit_sign_in(
     """An implicit login happened."""
     from bacommon.login import LoginType
 
-    _babase.app.accounts.on_implicit_sign_in(
+    assert _babase.app.plus is not None
+    _babase.app.plus.accounts.on_implicit_sign_in(
         login_type=LoginType(login_type_str),
         login_id=login_id,
         display_name=display_name,
@@ -362,7 +364,8 @@ def implicit_sign_out(login_type_str: str) -> None:
     """An implicit logout happened."""
     from bacommon.login import LoginType
 
-    _babase.app.accounts.on_implicit_sign_out(
+    assert _babase.app.plus is not None
+    _babase.app.plus.accounts.on_implicit_sign_out(
         login_type=LoginType(login_type_str)
     )
 
@@ -378,7 +381,8 @@ def login_adapter_get_sign_in_token_response(
     attempt_id = int(attempt_id_str)
     result = None if result_str == '' else result_str
 
-    adapter = _babase.app.accounts.login_adapters[login_type]
+    assert _babase.app.plus is not None
+    adapter = _babase.app.plus.accounts.login_adapters[login_type]
     assert isinstance(adapter, LoginAdapterNative)
     adapter.on_sign_in_complete(attempt_id=attempt_id, result=result)